Add Fuji F-Log / F-Gamut
Add Fuji F-Log / F-Gamut transforms for
Fuji cameras such as the X-T3. Includes:
 * Fuji F-Log transfer function
 * Fuji F-Log / F-Gamut transform to
    REC.709 scene linear from encoded
    footage

Integrated into the Filmic configuration
for simplicity sake, and the fact that
a majority of folks are most interested
in REC.709. The transforms are based
off of the offical documentation located
at:
https://www.fujifilm.com/support/digital_cameras/software/lut/pdf/F-Log_DataSheet_E_Ver.1.0.pdf

The F-Gamut encoding matches the BT.2020
pimaries, and as such, the matrix matches.
Given that BT.2020 has the same white point
chromaticity as the target working reference
space REC.709, there is no white point adaptation
required.